+++ Runtime/InternalBridge/AssemblyInfo.cs
@@ -1,4 +1,0 @@
-﻿using System.Runtime.CompilerServices;
-
-[assembly: InternalsVisibleTo("Unity.VectorGraphics")]
-[assembly: InternalsVisibleTo("Unity.VectorGraphics.Editor")]